+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Deezer Changelog
   * Fix track title and artist in queue view. Issue: tiliado/nuvolaruntime#609
   * Disable broken volume management.
   * Fix repeat button.
+  * Fix love button in queue view.
 
 3.2 - February 24th, 2019
 -------------------------